server/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
eventlet==0.39.0
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
gotrue==2.11.3
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.63.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
postgrest==0.19.3
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
realtime==2.3.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tqdm==4.67.1
typing_extensions==4.12.2
virtualenv==20.29.2
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
